Practice-02-2026/requirements.txt
beautifulsoup4==4.13.5
blinker==1.9.0
boto3==1.42.39
botocore==1.42.39
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.3.1
Flask==3.1.2
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
jmespath==1.1.0
MarkupSafe==3.0.3
numpy==2.3.3
pandas==2.3.2
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytz==2025.2
requests==2.32.5
s3transfer==0.16.0
six==1.17.0
soupsieve==2.8
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
Werkzeug==3.1.5
